pyproject.toml
[tool.poetry]
name = "kinnutils"
version = "0.1.0"
description = ""
authors = ["Ryan Farrell <rfarr@zestbloom.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-algorand-sdk = "2.0.0"
structlog = "^22.3.0"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
pydantic = "^1.10.5"
python-magic = "^0.4.27"
multihash = "^0.1.1"
six = "^1.16.0"
py-cid = "^0.3.0"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kinnutil = 'kinnutils.scripts.kinnutil:cli'